Refonce des fichiers de styles et rebuilt des composants modal de creation et d'edition des articles et de modification du profil utilisateur

+++ app/Http/Livewire/ProductEditor.php
@@ -2,11 +2,13 @@
 
 namespace App\Http\Livewire;
 
+use App\Models\User;
 use App\Models\Image;
 use App\Models\Product;
 use Livewire\Component;
 use Illuminate\Support\Str;
 use Livewire\WithFileUploads;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Storage;
 
 class ProductEditor extends Component
@@ -39,32 +41,35 @@ class ProductEditor extends Component
 
     public function updateProductGalery()
     {
-        $this->validate(['product_image' => 'image|max:1500|mimes:png,jpg,jpeg']);
-        $this->photoExtension = $this->product_image->extension();
-        $this->setImageName($this->photoExtension);
-        $first = null;
+        
+        if($this->authenticated()){
+            $this->validate(['product_image' => 'image|max:1500|mimes:png,jpg,jpeg']);
+            $this->photoExtension = $this->product_image->extension();
+            $this->setImageName($this->photoExtension);
+            $first = null;
 
-        $oldsImages = $this->product->images;
-        if($oldsImages->count() > 2){
-            $first = $oldsImages->first();
-        }
-        $this->product_image->storeAs('articlesImages', $this->getImageName());
-        $intoDB = Image::create(['name' => $this->getImageName(), 'product_id' => $this->product->id]);
-        if ($intoDB) {
-            if($first){
-                $local = Storage::delete($first->name);
-                $first->delete();
+            $oldsImages = $this->product->images;
+            if($oldsImages->count() > 2){
+                $first = $oldsImages->first();
             }
-            $this->dispatchBrowserEvent('hide-form');
+            $this->product_image->storeAs('articlesImages', $this->getImageName());
+            $intoDB = Image::create(['name' => $this->getImageName(), 'product_id' => $this->product->id]);
+            if ($intoDB) {
+                if($first){
+                    $local = Storage::delete($first->name);
+                    $first->delete();
+                }
+                $this->dispatchBrowserEvent('hide-form');
+                $this->emit('updatingFinish', true);
+                $this->dispatchBrowserEvent('FireAlert', ['title' => 'Mise à jour réussie', 'message' => "La mise à jour de la galerie s'est bien déroulée", 'type' => 'success']);
+            }
+            else{
+                $local = Storage::delete($this->getImageName());
+                $this->emit('updatingFinish', true);
+                $this->dispatchBrowserEvent('FireAlert', ['title' => 'Ereur serveur', 'message' => "La mise à jour de la galerie a échoué, veuillez réessayer!", 'type' => 'error']);
+            }
             $this->emit('updatingFinish', true);
-            $this->dispatchBrowserEvent('FireAlert', ['title' => 'Mise à jour réussie', 'message' => "La mise à jour de la galerie s'est bien déroulée", 'type' => 'success']);
         }
-        else{
-            $local = Storage::delete($this->getImageName());
-            $this->emit('updatingFinish', true);
-            $this->dispatchBrowserEvent('FireAlert', ['title' => 'Ereur serveur', 'message' => "La mise à jour de la galerie a échoué, veuillez réessayer!", 'type' => 'error']);
-        }
-        $this->emit('updatingFinish', true);
 
     }
 
@@ -86,6 +91,22 @@ class ProductEditor extends Component
         $product = Product::find($product_id);
         if($product){
             $this->emit('productUpdated', $product_id);
+        }
+    }
+
+
+    public function authenticated()
+    {
+        if(Auth::user()){
+            if(User::find(Auth::user()->id)->__hasAdminAuthorization()){
+                return true;
+            }
+            else{
+                return $this->dispatchBrowserEvent('FireAlertDoNotClose', ['title' => 'Authentification requise', 'message' => "Veuillez vous enthentifier avant de lancer des mises à jour", 'type' => 'warning']);
+            }
+        }
+        else{
+            return redirect()->route('login');
         }
     }
